refactor(report): conform to code style — extract utils, constants, types
No behavior change. Enforce the engine/module conventions across the
report feature:
- Pure/parsing functions moved out of services into .util.ts
  (verdict parsing, custom-id parsing, confidence bar, blockquote,
  Discord alert payload builder, review-url builder)
- Stray service constants grouped into .constant.ts with
  intent-revealing names
- Exported types moved to .type.ts
- De-duplicate requireUserId (report + review routers) into a shared
  engine util, renamed requireUserIdOrThrow (it throws)
- De-duplicate the email→display-name derivation (report + contributor
  services) into a shared engine util
- Drop leftover debug logging from the interactions controller

Discord message and OpenRouter services are now thin (DI + transport);
construction/parsing logic lives in utils.
