+++ packages/macgamingdb-server/src/engine/api/trpc/session-context.type.ts
@@ -1,0 +1,1 @@
+export type SessionContext = { user?: { user?: { id?: string } } | null };
