+++ packages/macgamingdb-server/src/engine/utils/derive-display-name-from-email.util.ts
@@ -1,0 +1,2 @@
+export const deriveDisplayNameFromEmail = (email: string): string =>
+  email.split('@')[0].replace(/[0-9._]/g, '');
