+++ packages/macgamingdb-server/src/modules/report/drivers/discord/types/discord-interaction-response.type.ts
@@ -1,0 +1,4 @@
+export type DiscordInteractionResponse = {
+  type: number;
+  data?: { content: string; components: [] };
+};
