+++ packages/macgamingdb-server/src/modules/report/drivers/discord/utils/format-blockquote.util.ts
@@ -1,0 +1,11 @@
+import { isNonEmptyString } from '@sniptt/guards';
+
+export const formatBlockquote = (text: string, fallback: string): string => {
+  if (!isNonEmptyString(text.trim())) {
+    return fallback;
+  }
+  return text
+    .split('\n')
+    .map((line) => `> ${line}`)
+    .join('\n');
+};
