+++ packages/macgamingdb-server/src/modules/report/drivers/discord/utils/parse-interaction-custom-id.util.ts
@@ -1,0 +1,14 @@
+type ParsedInteractionCustomId = { action: string; reviewId: string };
+
+export const parseInteractionCustomId = (
+  customId: string,
+): ParsedInteractionCustomId => {
+  const separatorIndex = customId.indexOf(':');
+  if (separatorIndex === -1) {
+    return { action: customId, reviewId: '' };
+  }
+  return {
+    action: customId.slice(0, separatorIndex),
+    reviewId: customId.slice(separatorIndex + 1),
+  };
+};
