+++ packages/macgamingdb-server/src/modules/report/drivers/openrouter/constants/openrouter-request.constant.ts
@@ -1,0 +1,5 @@
+export const OPENROUTER_COMPLETIONS_URL =
+  'https://openrouter.ai/api/v1/chat/completions';
+export const DEFAULT_MODERATION_MODEL = 'anthropic/claude-3.5-sonnet';
+export const MODERATION_MAX_ATTEMPTS = 2;
+export const WEB_SEARCH_MAX_RESULTS = 3;
