+++ packages/macgamingdb-server/src/modules/report/drivers/openrouter/services/openrouter-moderation.service.ts
@@ -1,35 +1,24 @@
 import { Injectable } from '@nestjs/common';
 import { z } from 'zod';
 import { isNonEmptyString } from '@sniptt/guards';
-import {
-  ModerationVerdictSchema,
-  type ModerationVerdict,
-} from '../../../dtos/moderation-verdict.dto';
+import { type ModerationVerdict } from '../../../dtos/moderation-verdict.dto';
 import { ReportException } from '../../../exceptions/report.exception';
 import { type JudgeReviewParams } from '../../../types/judge-review-params.type';
 import { type ModerationLlm } from '../../../types/moderation-llm.type';
+import {
+  DEFAULT_MODERATION_MODEL,
+  MODERATION_MAX_ATTEMPTS,
+  OPENROUTER_COMPLETIONS_URL,
+  WEB_SEARCH_MAX_RESULTS,
+} from '../constants/openrouter-request.constant';
 import { buildModerationPrompt } from '../utils/build-moderation-prompt.util';
-
-const OPENROUTER_COMPLETIONS_URL =
-  'https://openrouter.ai/api/v1/chat/completions';
-const DEFAULT_MODEL = 'anthropic/claude-3.5-sonnet';
-const MAX_ATTEMPTS = 2;
-const WEB_MAX_RESULTS = 3;
+import { parseModerationVerdict } from '../utils/parse-moderation-verdict.util';
 
 const completionResponseSchema = z.object({
   choices: z
     .array(z.object({ message: z.object({ content: z.string() }) }))
     .min(1),
 });
-
-const extractJsonObject = (content: string): string | undefined => {
-  const start = content.indexOf('{');
-  const end = content.lastIndexOf('}');
-  if (start === -1 || end === -1 || end < start) {
-    return undefined;
-  }
-  return content.slice(start, end + 1);
-};
 
 @Injectable()
 export class OpenRouterModerationService implements ModerationLlm {
@@ -42,11 +31,11 @@ export class OpenRouterModerationService implements ModerationLlm {
       );
     }
 
-    const model = process.env.MODERATION_MODEL ?? DEFAULT_MODEL;
+    const model = process.env.MODERATION_MODEL ?? DEFAULT_MODERATION_MODEL;
 
-    for (let attempt = 0; attempt < MAX_ATTEMPTS; attempt++) {
+    for (let attempt = 0; attempt < MODERATION_MAX_ATTEMPTS; attempt++) {
       const content = await this.requestCompletion({ apiKey, model, params });
-      const verdict = this.parseVerdict(content);
+      const verdict = parseModerationVerdict(content);
       if (verdict) {
         return verdict;
       }
@@ -74,7 +63,7 @@ export class OpenRouterModerationService implements ModerationLlm {
         temperature: 0,
         response_format: { type: 'json_object' },
         messages: buildModerationPrompt(params.params),
-        plugins: [{ id: 'web', max_results: WEB_MAX_RESULTS }],
+        plugins: [{ id: 'web', max_results: WEB_SEARCH_MAX_RESULTS }],
       }),
     });
 
@@ -94,19 +83,5 @@ export class OpenRouterModerationService implements ModerationLlm {
     }
 
     return parsed.data.choices[0].message.content;
-  }
-
-  private parseVerdict(content: string): ModerationVerdict | undefined {
-    const jsonText = extractJsonObject(content);
-    if (!isNonEmptyString(jsonText)) {
-      return undefined;
-    }
-
-    try {
-      const verdict = ModerationVerdictSchema.safeParse(JSON.parse(jsonText));
-      return verdict.success ? verdict.data : undefined;
-    } catch {
-      return undefined;
-    }
   }
 }
